added backup folder functionality to bash script, along with config file deletion & copying.

+++ install.sh
@@ -10,12 +10,29 @@ done
 DIR="$( cd -P "$( dirname "$SOURCE" )" && pwd )"
 echo "project directory: $DIR"
 #--
+read -p "will delete previous backup folder (if it exists). continue? y/n" $response
+if [[ $response = "y"]]
+then
+    rm -rf ~/.backups
+    echo "copying previous config files to ~/.backups/"
+    mkdir -p ~/.backups
+    mv ~/.emacs ~/.backups/.emacs
+    mv ~/.config/ ~/.backups/.config/
+    mv ~/.hyper.js ~/.backups/.emacs
 
-echo "linking emacs to ~/.emacs"
-ln -s "$DIR/emacs" ~/.emacs
+    echo "deleting previous config files..."
+    rm -rf ~/.emacs
+    rm -rf ~/.config/
+    rm -rf ~/.hyper.js
 
-echo "linking karabiner to ~/.config/karabiner"
-ln -s "$DIR/karabiner" ~/.config/
+    echo "beginning linking operation"
+    echo "linking emacs to ~/.emacs"
+    ln -s "$DIR/emacs" ~/.emacs
 
-echo "linking hyper.js to ~/.hyper.js"
-ln -s "$DIR/hyper.js" ~/.hyper.js
+    echo "linking karabiner to ~/.config/karabiner"
+    ln -s "$DIR/karabiner" ~/.config/
+
+    echo "linking hyper.js to ~/.hyper.js"
+    ln -s "$DIR/hyper.js" ~/.hyper.js
+else
+    echo "please manually move your files :)"
